+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## unreleased
 
-- Generate - Added option to generate operationId using template structures
+- Generate - Added option to generate operationId using template structures (#66)
+- CLI - Added option to bundle all options in --configFile
+- CLI - Added option to bundle all options in .openapiformatrc
 
 ## [1.22.5] - 2024-09-06